Fix problems introduced within last checkins.
git-svn-id: http://svn.ipd.uka.de/repos/ParaStation/trunk/mgmt@5876 34291a86-62e8-0310-8159-f060f8a2ee4e

+++ lib/psi/psi.c
@@ -507,8 +507,8 @@ int PSI_release(PStask_ID_t tid)
     }
 
     if (ret != msg.header.len || ret != sizeof(msg)) {
-	PSI_log(-1, "%s: PSI_recvMsg() got just %d/%d bytes (%d expected)\n",
-		__func__, ret, msg.header.len, sizeof(msg));
+	PSI_log(-1, "%s: PSI_recvMsg() got just %d/%d bytes (%ld expected)\n",
+		__func__, ret, msg.header.len, (long)sizeof(msg));
     }
 
     if (msg.header.type != PSP_CD_RELEASERES) {
